Add validates_associated to activemodel validation
If you're using activemodel, it will add 'validates_associated' to any
entity that requires activemodel validations. This is because in the
document/entity model we've defined, the ability to validate
associations is a central piece not included by activemodel (only
activerecord).

Also, namespaced previous tests and added a rake task to run all tests
as part of testing the new code.