Integrate jstore directly into Rails serialization
Previously, you were supposed to pass a block to ::jstore, which assumed
a 'jstore' column that was serialized as a hash. This use case allowed
for a single unstructured column as a bag to put any jstore 'entity' in,
and did not care about the details of the 'jstore' column beyond that.

Now, the default use case is to use ::jstore more like Rails' ::store,
where you pass the column name as the first argument. Unlike Rails'
store, however, it is assumed to be working with a 'document' object
(previously an 'entity'). Thus, the second argument to jstore is the
options hash, which configures things like the document class name,
namespace, or which column to use as the source of the document.

Note that the last option, called 'in', enables the previous use case of
storing many documents in a single column. Still, though, this column
needs to implement its own serialization, but jstore now ships with
JStore::JSONSerializer if you don't have a preferred json serializer.