Fix error when loading a nil document source
Previously, if a document source was nil, we would attempt to parse it
and choke. It's unclear what exactly the best thing to return is, ex. in
the collection case, should we return nil or [].

However, it's very clear that it shouldn't choke, so going with the
simplest answer for now and returning nil if document source is nil.