Enable proper nested attribute forms in rails
Previously, Rails forms with nested attributes happened to work with
singular relationships, but not entirely automatically (you had to
specify the relationship object). Collection-based relationships didn't
work at all.

This commit adds out-of-the-box compatability for nested relationships,
as it currently seems safe to assume that documents will be
handled almost entirely as nested types.

The problem was less well defined, so this commit does a bit more than
most:

* Defines "#[relationship]_attributes" and "#[relationship]_attributes="
  for all sub-document declarations or when using the :in option
* Changes the previous behavior of deep symbolizing document sources to
  deep stringifying document sources, as this both plays much nicer with
  Rails views and prevents memory bloat issues if we were to instead
  dynamically symbolize hashes from the view (note, with_indifferent_access
  was considered, but won't work with our pass-by-reference document
  instantiation)