Deserialize to an empty object when source is nil
Previously, the activerecord integration for a document returned nil
when source was nil. For example:

    class Blog < ActiveRecord
      dstore :author
    end

    Blog.last.author # nil if the column is nil in the database

This is reasonable in theory, but in practice it's nice to have an empty
document instead. This commit makes it return ex. an empty Blog::Author
instead of nil.

Note that since actual documents and ActiveRecord::dstore using
the :in option both have this empty-document behavior, this commit also
makes ActiveRecord#dstore (without :in option) consistent with the rest
of the featureset.